import from testazure fixes
