+++ hi-ml-azure/testazure/testazure/util.py
@@ -13,10 +13,8 @@ from contextlib import contextmanager
 from pathlib import Path
 from typing import Dict, Generator, Optional
 
-from azureml.core import Workspace
-
 from health_azure.utils import (ENV_RESOURCE_GROUP, ENV_SUBSCRIPTION_ID, ENV_WORKSPACE_NAME, WORKSPACE_CONFIG_JSON,
-                                _find_file, get_authentication, get_secret_from_environment)
+                                WorkspaceWrapper)
 
 DEFAULT_DATASTORE = "himldatasets"
 FALLBACK_SINGLE_RUN = "refs_pull_545_merge:refs_pull_545_merge_1626538212_d2b07afd"
@@ -48,45 +46,6 @@ def repository_root() -> Path:
     Gets the root folder of the git repository.
     """
     return himl_azure_root().parent
-
-
-def default_aml_workspace() -> Workspace:
-    """
-    Gets the default AzureML workspace that is used for testing.
-    """
-    config_json = _find_file(WORKSPACE_CONFIG_JSON)
-    if config_json is not None:
-        return Workspace.from_config(path=str(config_json))
-    else:
-        workspace_name = get_secret_from_environment(ENV_WORKSPACE_NAME, allow_missing=False)
-        subscription_id = get_secret_from_environment(ENV_SUBSCRIPTION_ID, allow_missing=False)
-        resource_group = get_secret_from_environment(ENV_RESOURCE_GROUP, allow_missing=False)
-        auth = get_authentication()
-        return Workspace.get(name=workspace_name,
-                             auth=auth,
-                             subscription_id=subscription_id,
-                             resource_group=resource_group)
-
-
-class WorkspaceWrapper:
-    """
-    Wrapper around aml_workspace so that it is lazily loaded, once.
-    """
-
-    def __init__(self) -> None:
-        """
-        Init.
-        """
-        self._workspace: Workspace = None
-
-    @property
-    def workspace(self) -> Workspace:
-        """
-        Lazily load the aml_workspace.
-        """
-        if self._workspace is None:
-            self._workspace = default_aml_workspace()
-        return self._workspace
 
 
 DEFAULT_WORKSPACE = WorkspaceWrapper()
