+++ hi-ml/testhiml/testhiml/utils/utils.py
@@ -1,0 +1,7 @@
+#  ------------------------------------------------------------------------------------------
+#  Copyright (c) Microsoft Corporation. All rights reserved.
+#  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+#  ------------------------------------------------------------------------------------------
+from health_azure.utils import WorkspaceWrapper
+
+DEFAULT_WORKSPACE = WorkspaceWrapper()
